feat: implement queue analysis with abandoned and timeout call statistics

+++ validation/userActivityFilters.ts
@@ -5,7 +5,7 @@ export const userActivityFiltersSchema = z
     fromDate: z.date(),
     toDate: z.date(),
     agents: z.array(z.object({ value: z.string(), label: z.string() })),
-    sla: z.number().min(1).max(100),
+    sla: z.number().int().min(1).max(100),
   })
   .refine(
     (data) => {
